types: migrate away from dataclass
This commits migrates away from dataclasses to custom overloaded
constructors. The main rationale behind doing so, is that, many
classes can take different types and number of params for initializing
an object. Based on the number of params, different C APIs can be called.
Also, the classes are not storing any native data. So, the data classes
are kind of ineffective.

Signed-off-by: Soumyadeep Ghosh <soumyadeepghosh2004@zohomail.in>